+++ test/ProcessEnvironment.cs
@@ -48,7 +48,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 
-namespace test
+namespace Shadowsocks.Test
 {
     static class ProcessEnvironment
     {
